+++ ea-commons/src/main/java/bluedot/electrochemistry/commons/entity/AlgorithmsLibraryApply.java
@@ -10,9 +10,9 @@ import java.time.LocalDateTime;
  */
 @Param
 public class AlgorithmsLibraryApply {
-    private Integer id;
-    private Integer userId;
-    private Integer algoId;
+    private Long id;
+    private Long userId;
+    private Long algoId;
 
     private String author;
 
@@ -29,27 +29,27 @@ public class AlgorithmsLibraryApply {
 
     private LocalDateTime gmt_modify;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public Integer getAlgoId() {
+    public Long getAlgoId() {
         return algoId;
     }
 
-    public void setAlgoId(Integer algoId) {
+    public void setAlgoId(Long algoId) {
         this.algoId = algoId;
     }
