+++ ea-commons/src/main/java/bluedot/electrochemistry/commons/entity/SystemFile.java
@@ -10,8 +10,8 @@ import java.time.LocalDateTime;
  */
 @Param
 public class SystemFile {
-    private Integer id;
-    private Integer userId;
+    private Long id;
+    private Long userId;
     /**
      * 记录此次备份的名称
      */
@@ -23,19 +23,19 @@ public class SystemFile {
     private String note;
     private LocalDateTime gmt_create;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
